#797: Split AdminUsersController into its own controller file
users.controller.ts mixed the user-facing UsersController and the
admin-only AdminUsersController in one file. Extract AdminUsersController
into admin-users.controller.ts for a clearer module boundary.
