+++ apps/backend/src/users/admin-users.controller.ts
@@ -1,0 +1,133 @@
+import {
+  Controller,
+  Get,
+  Param,
+  Query,
+  Patch,
+  Delete,
+  Post,
+  Body,
+  UseGuards,
+  Request,
+  NotFoundException,
+  UploadedFile,
+  UseInterceptors,
+} from '@nestjs/common';
+import { FileInterceptor } from '@nestjs/platform-express';
+import { UsersService } from './users.service';
+import { ApiBearerAuth, ApiTags, ApiOperation, ApiResponse, ApiBody, ApiConsumes } from '@nestjs/swagger';
+import { JwtAuthGuard } from '../auth/jwt-auth.guard';
+import { RolesGuard } from '../auth/roles.guard';
+import { Roles } from '../auth/roles.decorator';
+
+@ApiTags('admin')
+@ApiBearerAuth()
+@Controller('admin/users')
+@UseGuards(JwtAuthGuard, RolesGuard)
+export class AdminUsersController {
+  constructor(private readonly usersService: UsersService) {}
+
+  @Get()
+  @Roles('admin')
+  @ApiOperation({ summary: 'Get all users with filtering and pagination' })
+  @ApiResponse({
+    status: 200,
+    description: 'Returns paginated users',
+    schema: {
+      example: {
+        data: { users: [], total: 0, page: 1, limit: 10 },
+        statusCode: 200,
+        timestamp: '2024-01-01T00:00:00.000Z',
+      },
+    },
+  })
+  @ApiResponse({ status: 401, description: 'Unauthorized' })
+  @ApiResponse({ status: 403, description: 'Forbidden - admin role required' })
+  findAll(
+    @Query('page') page?: number,
+    @Query('limit') limit?: number,
+    @Query('role') role?: string,
+    @Query('isVerified') isVerified?: string,
+    @Query('search') search?: string
+  ) {
+    return this.usersService.findAll({
+      page: page ? Number(page) : 1,
+      limit: limit ? Number(limit) : 10,
+      role,
+      isVerified: isVerified === 'true' ? true : isVerified === 'false' ? false : undefined,
+      search,
+    });
+  }
+
+  @Post('import/csv')
+  @Roles('admin')
+  @ApiOperation({ summary: 'Bulk import users from a CSV file' })
+  @ApiConsumes('multipart/form-data')
+  @ApiBody({
+    schema: {
+      type: 'object',
+      properties: {
+        file: { type: 'string', format: 'binary' },
+      },
+    },
+  })
+  @UseInterceptors(FileInterceptor('file', { limits: { fileSize: 20 * 1024 * 1024 } }))
+  importUsers(@UploadedFile() file: Express.Multer.File, @Request() req: { user: { id: string } }) {
+    if (!file) {
+      throw new NotFoundException('CSV file is required for bulk import');
+    }
+    return this.usersService.bulkImportUsersCsv(file.buffer, req.user.id);
+  }
+
+  @Get('import-jobs/:jobId')
+  @Roles('admin')
+  @ApiOperation({ summary: 'Get bulk user import job status' })
+  getUserImportJob(@Param('jobId') jobId: string) {
+    return this.usersService.findImportJob(jobId);
+  }
+
+  @Patch(':id/role')
+  @Roles('admin')
+  @ApiOperation({ summary: 'Change user role' })
+  @ApiResponse({
+    status: 200,
+    description: 'Role updated successfully',
+    schema: { example: { data: {}, statusCode: 200, timestamp: '2024-01-01T00:00:00.000Z' } },
+  })
+  @ApiResponse({ status: 401, description: 'Unauthorized' })
+  @ApiResponse({ status: 403, description: 'Forbidden - admin role required' })
+  @ApiResponse({ status: 404, description: 'User not found' })
+  changeRole(@Param('id') id: string, @Body('role') role: string) {
+    return this.usersService.changeRole(id, role);
+  }
+
+  @Patch(':id/ban')
+  @Roles('admin')
+  @ApiOperation({ summary: 'Ban or unban a user' })
+  @ApiResponse({
+    status: 200,
+    description: 'User ban status updated',
+    schema: { example: { data: {}, statusCode: 200, timestamp: '2024-01-01T00:00:00.000Z' } },
+  })
+  @ApiResponse({ status: 401, description: 'Unauthorized' })
+  @ApiResponse({ status: 403, description: 'Forbidden - admin role required' })
+  @ApiResponse({ status: 404, description: 'User not found' })
+  banUser(@Param('id') id: string, @Body('isBanned') isBanned: boolean) {
+    return this.usersService.banUser(id, isBanned);
+  }
+
+  @Delete(':id')
+  @Roles('admin')
+  @ApiOperation({ summary: 'Soft delete a user' })
+  @ApiResponse({
+    status: 200,
+    description: 'User deleted successfully',
+    schema: { example: { data: {}, statusCode: 200, timestamp: '2024-01-01T00:00:00.000Z' } },
+  })
+  @ApiResponse({ status: 401, description: 'Unauthorized' })
+  @ApiResponse({ status: 403, description: 'Forbidden - admin role required' })
+  @ApiResponse({ status: 404, description: 'User not found' })
+  deleteUser(@Param('id') id: string) {
+    return this.usersService.softDelete(id);
+  }
+}
