+++ apps/backend/src/users/users.module.ts
@@ -3,7 +3,8 @@ import { TypeOrmModule } from '@nestjs/typeorm';
 import { ScheduleModule } from '@nestjs/schedule';
 import { User } from './user.entity';
 import { UsersService } from './users.service';
-import { UsersController, AdminUsersController } from './users.controller';
+import { UsersController } from './users.controller';
+import { AdminUsersController } from './admin-users.controller';
 import { StellarModule } from '../stellar/stellar.module';
 import { ImportJob } from '../import-export/import-job.entity';
